+++ changelog.md
@@ -1,6 +1,7 @@
 # Change Log
 
 ## `head`
+- enable partitionKey for sendBatch
 
 ## `v2.0.1`
 - update to amqp 0.11.2 & common 2.1.0 to fix [#115](https://github.com/Azure/azure-event-hubs-go/issues/115)