+++ changelog.md
@@ -1,7 +1,9 @@
 # Change Log
 
 ## `head`
-- enable partitionKey for sendBatch
+
+## `v2.0.2`
+- enable partitionKey for sendBatch to fix [#128](https://github.com/Azure/azure-event-hubs-go/issues/128)
 - ensure sender receives ack'd messages from EH [#126](https://github.com/Azure/azure-event-hubs-go/issues/126)
 - close `leaseCh` on function return in storage.(*LeaserCheckpointer).GetLeases to fix [#136](https://github.com/Azure/azure-event-hubs-go/issues/136)